check for utilities before making unite mappings
Some unite functionality depends on system tools: unite bibtex needs pybtex and a specified bib file; unite grep requires a grep-like utility like highway, ack-grep, silver searcher of platinum searcher; and unite find on Windows needs a find-like tool like silver searcher or platinum searcher. For unite commands that depend on system utilities the mappings are first made to a function which, when the mapping is invoked, checks that the required system tools are available. If the tools are available the user is provided with an informative message, the mapping is made to the unite command for future invocations, and the unite command is then invoked. If the tools are not available the user is provided with an informative error message and the mapping is left unchanged.